requirements.txt
-e git+https://github.com/django/django.git@ad96254af98f961d6691481c27127613afeee840#egg=Django-dev
Scrapy==0.24.4
Twisted==14.0.0
Unipath==1.0
cffi==0.8.6
characteristic==14.1.0
click==3.3
cryptography==0.5.4
cssselect==0.9.1
dj-database-url==0.3.0
dj-static==0.0.6
django-localflavor-br==1.0
django-toolbelt==0.0.1
gnureadline==6.3.3
gunicorn==19.1.1
ipdb==0.8
ipython==2.2.0
lxml==3.3.5
plan==0.4
psycopg2==2.5.3
pyOpenSSL==0.14
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.10
python-decouple==2.3
queuelib==1.2.2
requests==2.3.0
service-identity==14.0.0
six==1.7.3
static3==0.5.1
w3lib==1.10.0
wsgiref==0.1.2
zope.interface==4.1.1
